+++ packages/openapi-types/src/openapi-types.ts
@@ -60,10 +60,7 @@ export namespace OpenAPI {
 
   export type SchemaObject = OpenAPIV2.SchemaObject | OpenAPIV3.SchemaObject | OpenAPIV3_1.SchemaObject
 
-  export type HttpMethod =
-    | keyof typeof OpenAPIV2.HttpMethods
-    | keyof typeof OpenAPIV3.HttpMethods
-    | OpenAPIV3_1.HttpMethods
+  export type HttpMethod = keyof typeof OpenAPIV2.HttpMethods | OpenAPIV3.HttpMethods | OpenAPIV3_1.HttpMethods
 }
 
 // biome-ignore lint/style/noNamespace: We want it to be a module here.
@@ -357,16 +354,7 @@ export namespace OpenAPIV3 {
   // See https://swagger.io/specification/#path-item-object
   // You can use keys or values from it in TypeScript code like this:
   //     for (const method of Object.values(OpenAPIV3.HttpMethods)) { … }
-  export enum HttpMethods {
-    GET = 'get',
-    PUT = 'put',
-    POST = 'post',
-    DELETE = 'delete',
-    OPTIONS = 'options',
-    HEAD = 'head',
-    PATCH = 'patch',
-    TRACE = 'trace',
-  }
+  export type HttpMethods = 'get' | 'put' | 'post' | 'delete' | 'options' | 'head' | 'patch' | 'trace' | 'connect'
 
   export type PathItemObject<T = {}> = {
     $ref?: string
@@ -573,14 +561,14 @@ export namespace OpenAPIV3 {
     description?: string
     scheme?: string
     bearerFormat?: string
-  }
+  } & AnyOtherAttribute
 
   export type ApiKeySecurityScheme = {
     type?: 'apiKey'
     description?: string
     name?: string
     in?: string
-  }
+  } & AnyOtherAttribute
 
   export type OAuth2SecurityScheme = {
     type?: 'oauth2'
@@ -590,23 +578,23 @@ export namespace OpenAPIV3 {
         authorizationUrl?: string
         refreshUrl?: string
         scopes?: { [scope: string]: string }
-      }
+      } & AnyOtherAttribute
       password?: {
         tokenUrl?: string
         refreshUrl?: string
         scopes?: { [scope: string]: string }
-      }
+      } & AnyOtherAttribute
       clientCredentials?: {
         tokenUrl?: string
         refreshUrl?: string
         scopes?: { [scope: string]: string }
-      }
+      } & AnyOtherAttribute
       authorizationCode?: {
         authorizationUrl?: string
         tokenUrl?: string
         refreshUrl?: string
         scopes?: { [scope: string]: string }
-      }
+      } & AnyOtherAttribute
     }
   }
 
@@ -614,7 +602,7 @@ export namespace OpenAPIV3 {
     type?: 'openIdConnect'
     description?: string
     openIdConnectUrl?: string
-  }
+  } & AnyOtherAttribute
 
   export type TagObject = {
     name?: string