+++ integrations/nextjs/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @scalar/nextjs-api-reference
 
+## 0.8.5
+
+### Patch Changes
+
+- 6ead9cd: chore: remove unused dependencies
+- 4156f1d: Expand workspace store integration
+  - @scalar/core@0.3.3
+
+## 0.8.4
+
+### Patch Changes
+
+- @scalar/core@0.3.2
+- @scalar/types@0.2.2
+
 ## 0.8.3
 
 ### Patch Changes