+++ packages/build-tooling/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @scalar/build-tooling
 
+## 0.2.3
+
+### Patch Changes
+
+- 4156f1d: Expand workspace store integration
+
+## 0.2.2
+
+### Patch Changes
+
+- 8165b3b: feat(helpers): added new helpers package
+
 ## 0.2.1
 
 ### Patch Changes